requirements.txt
streamlit==1.32.0
pandas==2.2.1
gspread==5.12.4
google-auth==2.27.0
matplotlib==3.8.0
